+++ src/index.js
@@ -1,39 +1,6 @@
 import './style.scss';
 import './style.css';
-
-const renderComment = (item) => {
-  console.log(item);
-  const body = document.querySelector('.container');
-  const modalBG = document.createElement('div');
-  modalBG.classList.add('modal');
-  const commentUI = document.createElement('div');
-  commentUI.classList.add('comment-container');
-  commentUI.innerHTML = `
-      <img src="${item.strMealThumb}" />
-      
-      <h2>${item.strMeal}</h2>
-    
-    `;
-
-  const details = document.createElement('div');
-  const closeButton = document.createElement('h2');
-  closeButton.innerHTML = 'X';
-  closeButton.classList.add('close');
-  closeButton.addEventListener('click', () => body.classList.toggle('open'));
-  details.classList.add('ingredient');
-  for (let i = 1; i < 5; i++) {
-    const p = document.createElement('p');
-    p.innerHTML = `${item['strIngredient' + i]} : ${item['strMeasure' + i]}`;
-    details.appendChild(p);
-  }
-
-  commentUI.appendChild(details);
-
-  modalBG.appendChild(commentUI);
-  modalBG.appendChild(closeButton);
-
-  body.appendChild(modalBG);
-};
+import renderComment from './modules/renderComment';
 
 const getComments = async (id) => {
   const response = await fetch(
